requirements.txt
aiohttp==3.9.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
amqp==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
billiard==4.2.0
blinker==1.7.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==7.0.0
dpath==2.1.6
dulwich==0.21.7
dvc==3.49.0
dvc-data==3.15.1
dvc-gdrive==3.0.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
entrypoints==0.4
executing==2.0.1
filelock==3.13.4
Flask==3.0.3
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
funcy==2.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
grandalf==0.8
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gto==1.7.1
httplib2==0.22.0
hydra-core==1.3.2
idna==3.7
imbalanced-learn==0.12.2
imblearn==0.0
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.23.0
iterative-telemetry==0.0.8
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
kombu==5.3.7
Mako==1.3.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==2.12.1
multidict==6.0.5
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
orjson==3.10.0
packaging==24.0
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.3.0
platformdirs==3.11.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydot==2.0.0
PyDrive2==1.19.0
pygit2==1.14.1
Pygments==2.17.2
pygtrie==2.5.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
requests==2.31.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.4.2
scipy==1.13.0
scmrepo==3.3.1
seaborn==0.13.2
semver==3.0.2
setuptools==69.5.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
sklearn-features==0.0.2
smmap==5.0.1
SQLAlchemy==2.0.29
sqlparse==0.5.0
sqltrie==0.11.0
stack-data==0.6.3
tabulate==0.9.0
threadpoolctl==3.4.0
tomlkit==0.12.4
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.3
voluptuous==0.14.2
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
xgboost==2.0.3
yarl==1.9.4
zc.lockfile==3.0.post1
zipp==3.18.1
